Extract BrokerConfigValidator.
